+++ src/steps/step2_3.py
@@ -3,16 +3,16 @@ Train two agents with your algorithm of choice, on the source and target domain 
 Then, test each model and report its average reward over 50 test episodes.
 In particular, report results for the following “training→test” configurations: source→source, source→target (lower bound), target→target (upper bound).
 """
-from functools import partial
 import os
+import shutil
+from functools import partial
 from typing import Callable
 
 import optuna
-import gym
-from model.env.custom_hopper import *
-from torch.utils.tensorboard import SummaryWriter
 from stable_baselines3 import SAC
-import shutil
+from torch.utils.tensorboard import SummaryWriter
+
+from model.env.custom_hopper import *
 
 
 def constant_schedule(initial_value: float) -> Callable[[float], float]:
@@ -77,6 +77,8 @@ def objective_fn(trial, logdir='.'):
                 model.learn(total_timesteps=50_000, progress_bar=True,
                             tb_log_name=f"SAC_training_{source}")
 
+                model.save(os.path.join("trained_models", f"step2_trial_{trial.number}_env_{source}"))
+
                 last_trained_env = source
 
             n_episodes = 50
@@ -107,13 +109,18 @@ def objective_fn(trial, logdir='.'):
     return src_trg_avg_return
 
 
-def main(base_prefix='.'):
+def main(base_prefix='.', force=False):
     logdir = f"{base_prefix}/sac_tb_step2_3_log"
 
-    try:
-        shutil.rmtree(logdir)
-    except Exception as e:
-        print(e)
+    if os.path.isdir(logdir):
+        if force:
+            try:
+                shutil.rmtree(logdir)
+            except Exception as e:
+                print(e)
+        else:
+            print(f"Directory {logdir} already exists. Shutting down...")
+            return
 
     search_space = {
         "gamma": [0.9, 0.99],
